# docs/swagger.yaml
basePath: /
definitions:
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      token:
        type: string
    type: object
  hr.CreateEmployeeRequest:
    properties:
      businessUnitId:
        type: string
      departmentId:
        type: string
      displayName:
        type: string
      employeeNo:
        type: string
      firstName:
        type: string
      jobTitleId:
        type: string
      lastName:
        type: string
      managerId:
        type: string
      workEmail:
        type: string
    required:
    - employeeNo
    - firstName
    - lastName
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Deep QMS backend application serving the core HR, IT, and specialized
    logic domains.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DML API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user via email and password and returns a JWT token
        for Authorization.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully authenticated
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad request payload
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
      summary: Login and get JWT token
      tags:
      - Authentication
...
  /api/v1/employees:
    get:
      description: Get a paginated list of employees.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: pageSize
        type: integer
      - description: Search fuzzy match
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated Employee data
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Employees
      tags:
      - Employees
    post:
      consumes:
      - application/json
      description: Adds a new employee to the human resources schema. Supports assigning
        to Organizational Units (Business Unit, Department, Job Title) and a Manager
        for tracking hierarchical reporting lines.
      parameters:
      - description: Employee Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hr.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create an Employee
      tags:
      - Employees
  /api/v1/employees/{id}:
    get:
      description: Fetch a single employee record.
      parameters:
      - description: Employee UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Employee by ID
      tags:
      - Employees
  /api/v1/employees/{id}/hierarchy:
    get:
      description: Fetches an employee and a recursively mapped tree of their direct
        and indirect reporting subordinates.
      parameters:
      - description: Employee UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Get Employee Hierarchy
      tags:
      - Employees
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"